+++ src/flytetest/config.py
@@ -93,6 +93,8 @@ AGAT_CONVERSION_WORKFLOW_NAME = "annotation_postprocess_agat_conversion"
 AGAT_CONVERSION_RESULTS_PREFIX = "agat_conversion_results"
 AGAT_CLEANUP_WORKFLOW_NAME = "annotation_postprocess_agat_cleanup"
 AGAT_CLEANUP_RESULTS_PREFIX = "agat_cleanup_results"
+TABLE2ASN_WORKFLOW_NAME = "annotation_postprocess_table2asn"
+TABLE2ASN_RESULTS_PREFIX = "table2asn_results"
 
 TASK_ENVIRONMENT_CONFIGS: tuple[TaskEnvironmentConfig, ...] = (
     TaskEnvironmentConfig(name=TASK_ENV_NAME),
@@ -122,6 +124,7 @@ TASK_ENVIRONMENT_CONFIGS: tuple[TaskEnvironmentConfig, ...] = (
     TaskEnvironmentConfig(name=AGAT_WORKFLOW_NAME),
     TaskEnvironmentConfig(name=AGAT_CONVERSION_WORKFLOW_NAME),
     TaskEnvironmentConfig(name=AGAT_CLEANUP_WORKFLOW_NAME),
+    TaskEnvironmentConfig(name=TABLE2ASN_WORKFLOW_NAME),
 )
 
 TASK_ENVIRONMENT_NAMES = tuple(config.name for config in TASK_ENVIRONMENT_CONFIGS)
@@ -147,6 +150,7 @@ eggnog_env = TASK_ENVIRONMENTS_BY_NAME[EGGNOG_WORKFLOW_NAME]
 agat_env = TASK_ENVIRONMENTS_BY_NAME[AGAT_WORKFLOW_NAME]
 agat_conversion_env = TASK_ENVIRONMENTS_BY_NAME[AGAT_CONVERSION_WORKFLOW_NAME]
 agat_cleanup_env = TASK_ENVIRONMENTS_BY_NAME[AGAT_CLEANUP_WORKFLOW_NAME]
+table2asn_env = TASK_ENVIRONMENTS_BY_NAME[TABLE2ASN_WORKFLOW_NAME]
 DEFAULT_SLURM_ACCOUNT = os.environ.get("FLYTETEST_SLURM_ACCOUNT", "rcc-staff")